Vehicle wyatt

+++ src/main/java/groupa/vehicle/Aircraft.java
@@ -3,9 +3,13 @@ package groupa.vehicle;
 import groupa.farm.Field;
 
 public abstract class Aircraft extends Vehicle{
-    public Aircraft(String sound, int id) {
+    public Aircraft(String sound, long id) {
         super(sound, id);
     }
 
+    /**
+     * will get overridden in cropduster class
+     * @param field
+     */
     public abstract void fly(Field field);
 }
