+++ src/main/java/groupa/vehicle/CropDuster.java
@@ -7,10 +7,14 @@ import groupa.farm.CropRow;
 import java.util.ArrayList;
 
 public class CropDuster extends Aircraft implements FarmVehicle {
-    public CropDuster(String sound, int id) {
+    public CropDuster(String sound, long id) {
         super(sound, id);
     }
 
+    /**
+     * fertilizes each crop row im a field
+     * @param field
+     */
     private void fertilize(Field field) {
     ArrayList<CropRow> cropRows = field.getCropRows();
     for (CropRow cropRow : cropRows) {
@@ -18,6 +22,11 @@ public class CropDuster extends Aircraft implements FarmVehicle {
     }
 
     }
+
+    /**
+     * calls fertilize and checks if cropduster is set in use
+     * @param field
+     */
     @Override
     public void fly(Field field){
         if (isInUse()) {
@@ -28,12 +37,20 @@ public class CropDuster extends Aircraft implements FarmVehicle {
             System.out.println("No pilot to fertilize field with crop duster!");
         }
     }
+
+    /**
+     * sets the cropduster to inuse
+     * @param inUse
+     */
     @Override
     public void operate(boolean inUse) {
        setInUse(inUse);
 
     }
 
+    /**
+     * plays sound
+     */
     @Override
     public void makeNoise() {
         System.out.println(getSound());
