+++ src/main/java/groupa/vehicle/Tractor.java
@@ -7,10 +7,15 @@ import java.util.ArrayList;
 
 public class Tractor extends Vehicle implements FarmVehicle {
 
-    public Tractor(String sound, int id) {
+    public Tractor(String sound, long id) {
         super(sound, id);
     }
 
+    /**
+     * The harvest method in the Tractor class is responsible for harvesting all the crops in a given field.
+     * It checks if the tractor is in use and then iterates over each crop row in the field, calling the harvestAllCrops method on each crop row
+     * @param field
+     */
     public void harvest(Field field){
 
         if (isInUse()) {
@@ -28,12 +33,18 @@ public class Tractor extends Vehicle implements FarmVehicle {
 
     }
 
-
+    /**
+     * this sets the tractor in use
+     * @param inUse
+     */
     @Override
     public void operate(boolean inUse) {
       setInUse(inUse);
     }
 
+    /**
+     * this plays the tractor sound
+     */
     @Override
     public void makeNoise() {
         System.out.println(getSound());
